[v2.0 samples] SDL "keycodes" with "scancodes"
In SDL2, Scancodes were introduced. This keyboard representation does not move around with the keyboard layout and makes everything consistant.

When you use a non-qwerty keyboard, the alpha keys tend to be moved around. Scancodes makes so that, for example, the WASD keys will allways be placed at the right place.

This is a non issue on Windows as it always behaved that way, but on linux, Keycodes respect the current keyboard layout, so international users will have the samples running with the wrong keyboard mapping.

All of the alphabetical keys in the samples source code have been changed to scancodes. I did not change all the function key and other special characters as this will not change anything for theses keys and I wanted to keep the diff short.