Fixed crash when trying to use AbsolutePixelCountLodStrategy
AbsolutePixelCountLodStrategy needs a valid Viewport, however
LOD calculations happen before the Viewport is setup.

This caused a catch-22 problem because setRenderPassDescToCurrent setups
the Viewport, however it can't be called before calculating LODs, but
LODs need setRenderPassDescToCurrent to be called.

Solution was to manually setup just a single Viewport in the LOD camera
so that LOD calculations can proceed, since the full call to
setRenderPassDescToCurrent isn't actually needed