Multiple shadow mapping improvements (#100). See description
Changes:
- Added HlmsMacroblock::mDepthClamp
- Added RSC_DEPTH_CLAMP, present almost anywhere except A7, A8 iOS HW
- Added shadow map pancacking using mDepthClamp to
FocusedShadowCameraSetup for directional shadow maps. This tightens
depth range, increasing precision and decreasing the need for 32-bit
depth buffers
- Fix normal offset bias heavily dependent on shadow map resolution
- Point lights' normal offset bias was accidentally hardcoded
- shadowConstantBias in the caster shader no longer varies implicitly
due to depth range. It was a cryptic (user hostile) way of increasing
bias to higher splits.
- Changed default value of constantBiasScale
- Changed default value of normalOffsetBias
- Added autoConstantBiasScale and autoNormalOffsetBiasScale. This value
adjusts the bias based on the size of the shadow camera. Shadow cameras
covering more area require bigger biases. This setting is a much
user-friendlier way of automatically scaling the bias to higher splits.