Fix bug with compositor texture referenced from TextureUnitState
When setting the reference to the compositor texture, mFrames[0] was set to the texture name.
However, this caused the TextureUnitState to try and read the file from disk when the
TextureUnitState was loaded.

Instead, mFrames[0] should not be set to anything. mFramePtrs[0] will be correctly set
when the corresponding compositor is updated.