Fix crash at shutdown in TextureUnitState
When the TextureUnitState reference a compositor texture, the TextureUnitState does not
add itself as a listener to the TextureGpu. This means that if the TextureGpu is then
destroyed before the TextureUnitState, the TextureUnitState will have a dangling reference
to the GpuTexture in mFramePtrs, and will try to remove itself as a listener from GpuTexture
at destruction, thus de-referencing an invalid pointer.