Fix Aabb::distance calculation
The old implementation worked like this:

1) Calculate the vector from v to the center of the Aabb
2) Update the distance by taking the absolute value of each element,
   and then subtracting the corresponding half size of the AABB
3) Return the minimum element of the updated distance,
   or 1 if the updated distance is less than 1

Disregarding the fact that the function will always return a positive value
even if the position v is inside the cube, the algorithm is still flawed.
For instance, imagine a position that is infinitely far away along the X axis:

v = { inf, 0.0, 0.0 }

And a AABB located at the center with half size 1 along each axis.

Then, the algorithm works as follows:

1) dist = { 0.0, 0.0, 0.0 } - { inf, 0.0, 0.0 } = { -inf, 0.0, 0.0 }
2) dist = abs(dist) - halfSize = abs({ -inf, 0.0, 0.0 }) - { 1.0, 1.0, 1.0 }
        = { inf, 0.0, 0.0 } - { 1.0, 1.0, 1.0 } = { inf, -1.0, -1.0 }
3) return max(min(dist), 1.0) = max(-1.0, 1.0) = 1.0

So given a point infinitely far away a finite AABB, we get a distance of 1.0.

This change updates the algorithm:

1) Calculate the vector from v to the center of the Aabb
2) Update the distance by taking the absolute value of each element,
   then subtract the corresponding half size of the AABB,
   and finally clamp the value to 0 if it ends up being negative
3) Return the length of the updated distance

In order to allow the user to avoid a square root, the algorithm is split in
two parts. The first part does the majority of the work, but returns the distance
squared. The second part calculates the square root.