Fix compiler errors
Using "private using" isn't a perfect solution because overload
resolution resolution happens before access control.

This triggers "ambiguous call error" when two functions use default
arguments in a way that it is impossible for the compiler to distinguish
which version is being used.

The only solution for these cases is to silence the compiler warning via
pragmas.