[Vulkan] Fix GPU deadlock/freeze
If VaoManager::_update would be called twice in a frame, it would cause
two calls to commitAndNextCommandBuffer one with
SubmissionType::NewFrameIdx, then a later one with
SubmissionType::EndFrameAndSwap

The 1st call would consume mGpuWaitSemaphForCurrCmdBuff and reset them.
The 2nd call would use mGpuWaitSemaphForCurrCmdBuff again (wrong) and
since they were reset, it would wait forever for a semaphore that will
never be signaled.

The fix is to clear mGpuWaitSemaphForCurrCmdBuff after we've sent them
to the GPU.

Thanks to user dermont for the assistance on this bug.
See https://forums.ogre3d.org/viewtopic.php?p=549492#p549492