Alpha tested textured shadows could be wrong (#163)
Fix #163

When alpha tested textured shadows were in the scene, it would often
look correct by accident. calculateHashForPreCaster was removing many
Hlms properties that were critical for sourcing from the correct texture
and sampler slot.

Additionally it was crashing in Vulkan because the number of declared
textures in the shader must match the number of textures in a
DescriptorSetTexture

The solution is to create those properties again