[Vulkan] Fix corruption when RenderTextures are destroyed (#221)
When a TextureGpu is destroyed and a new TextureGpu is given the same
address ptr (or when a TextureGpu is no longer Resident, then Resident
again) VulkanRenderPassDescriptor::calculateSharedKey will look at its
cache, see that there is already an entry but it will contain old
values.

Then VulkanRenderPassDescriptor::setupFbo sees fboDesc.mRenderPass is
already initialized; when in fact fboDesc.mRenderPass should be
destroyed and reinitialized.