Fix ParticleFX rendering bugs during the first frame
An unfortunate chain of events would lead to calculateHashFor delaying
hash calculation but it is already too late.

The problem is that BillboardSet has that _createBuffers hack and should
not set the material inside _updateRenderQueue.

The solution (workaround) is to delay rendering of the PFX to the next
frame if it's not yet ready.