[Vk] AsyncTextureTicket may use the wrong image aspect (#322)
VulkanAsyncTextureTicket was using mPixelFormatFamily but PFG_D32_FLOAT
belongs to PFG_R32_FLOAT family. The former is DEPTH, the latter is
COLOR.

Use the pixel format of the texture rather than the formaty family to
select the right image aspect.

Additionally StagingTexture was always wrongly assuming the image aspect
is COLOR