Do not treat dGPUs as UMA
When the BIOS & GPU has Resizable BAR enabled, x86 GPUs will expose
DEVICE_LOCAL_BIT + HOST_VISIBLE_BIT + HOST_COHERENT_BIT memory.

That doesn't mean they're UMA, and we definitely do NOT want to mess
with the can of worms that is dealing with that memory type.

And if Resizable BAR is disabled, that memory type may still be exposed
but limited to 256MB (which is globally shared by the entire OS, if the
OS runs out of it bad things happen).

There's still the question of what to do with AMD iGPUs though (e.g.
Vega and RDNA2 iGPUs will expose these flags and we will consider them
UMAs. Technically this is correct. Is it a good idea? I don't know).