Remove most references to GLES and GLSLES
Eugene is not going to be pursuing this endeavour (he is the one who
created the GLES/GLSLES split) anymore.

Furthermore it is likely that the Hlms should reuse GLSL code and use a
few extra property to mark the GLSLES differences.

It also doesn't make sense to have fully separate files just to write

precision highp float;

When all of our GLSL shaders now start with "#version ogre_glsl_ver_330"
to patch this from C++.