Fix uninitialized texture in Sample_LocalCubemapsManualProbes
When using PCC (the original version, not the Auto/per-pixel one) and
manually assigning cubemaps to materials the routine
CubemapProbe::_clearCubemap would clear the wrong texture.

It would clear the temporary RTT (which is used to avoid a GPU feedback
loop i.e. rendering and sampling to/from same texture) instead of the
real cubemap texture.

The still-uninitialized cubemap texture would then be copied to the
proxy PCC cubemap; and then the temporary RTT cubemap would be rendered
normally, using the uninitialized proxy PCC; causing artifacts.

If Compute Shaders are supported, the cubemap texture must be cleared
using Cubemap textures, which is why now Compute/Tools folder was added
to resources2.cfg file so the necessary Compute Shaders are available.