[Vk] Fix validation VUID-vkQueueSubmit-pSignalSemaphores-00067
This validation is new to Vulkan SDK 1.4.313.0 and hit a lot of apps.

The problem is that we assume that after N frames, the semaphores should
be available again. This may not be true because the presentation engine
is not required to release them in the order we expect (even in FIFO
mode).

So when we do:

```cpp
waitForTailFrameToFinish();

FastArray<UsedSemaphore>::iterator itor = mUsedSemaphores.begin();
FastArray<UsedSemaphore>::iterator endt = mUsedSemaphores.end();

while( itor != endt && ( mFrameCount - itor->frame ) >=
mDynamicBufferMultiplier )
{
    mAvailableSemaphores.push_back( itor->semaphore );
    ++itor;
}

mUsedSemaphores.erasePOD( mUsedSemaphores.begin(), itor );
```

We put hazardous semaphores into mAvailableSemaphores.

The solution is to track these semaphores separately,  and don't recycle
them until we receive confirmation via vkAcquireNextImageKHR's
swapchainIdx that the swapchain has been released.