Fix setTextures() & setSamplerblock() sometimes not taking effect
The calculation of needsRecalculateHash is even more complex than
initially thought.

Consider the following case:

1. textureMaps[2]
2. Diffuse samples from textureMaps[0]
3. Specular samples from textureMaps[1]

Afterwards the texture arrays are swapped:

1. textureMaps[2]
2. Diffuse samples from textureMaps[1]
3. Specular samples from textureMaps[0]

We would not detect this case because the old textureMaps[] and the new
one are the same.

The same problem happens with samplers, where it's much more likely to
happen because samplerStates[] could contain 3 samplers and how are they
assigned to each texture unit could change, yet the baked descriptor
array would not change.

The solution is to always assume hash must be recalculated if textures
or samplers become dirty. This is simpler than performing an even more
complex search to see if we should recalculate the hashes and call
flushRenderables().