[Apple] Remove OpenGL3+ support
Metal won't compile due to a compiler error (see
https://forums.ogre3d.org/viewtopic.php?p=557402&sid=15e306ceafb9e4328786e729251251cc#p557402)
in XCode's SDK if OpenGL is found.

Apple keeps making it more difficult to support OpenGL because clearly
they don't test it thoroughly nor are interested.
We don't have enough manpower to fix every little thing that breaks on
Apple because of OpenGL.