Fix recursive calling of CompositorNode::notifyRecreated
Fixes #539
Closes #555

The whole point of CompositorNode::mConnectedNodes was to prevent
unnecessary updates to unrelated nodes, which could lead to O(N²)
behavior.

However due to implementation issues, the "solution" was causing
something similar to 2^N behavior or worse, which is much worse than N².

The problem is that mConnectedNodes could contain the same node twice or
more; and if a texture was found as an output, the same node would be
called N times (which could cascade into being called M times, then O
times...).

This solution does not affect ABI, thus it should be possible to
backport to older versions.

Kudos to user https://github.com/nimble0 who found the problem and
proposed a solution.